# docs/api/openapi.yaml
components:
  schemas:
...
    ClassificationHistoryResponse:
      description: Historical classification results for a document.
      properties:
        classifications:
          items:
            $ref: '#/components/schemas/DocumentClassificationResponse'
          title: Classifications
          type: array
        document_id:
          format: uuid
          title: Document Id
          type: string
        document_type:
          title: Document Type
          type: string
        total:
          title: Total
          type: integer
      required:
      - document_id
      - document_type
      - classifications
      - total
      title: ClassificationHistoryResponse
      type: object
    DocumentClassificationRequest:
      description: 'Request to classify a document into taxonomy and topics.


        The user provides the document content and optional constraints.

        The system will suggest taxonomy first, then topics within that taxonomy.'
      properties:
        content:
          description: Document content to classify
          maxLength: 50000
          minLength: 1
          title: Content
          type: string
        document_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Optional document ID if classifying existing document
          title: Document Id
        document_type:
          anyOf:
          - type: string
          - type: 'null'
          description: Type of document (Note, Document, etc.)
          title: Document Type
        max_topics:
          default: 5
          description: Maximum number of topics to suggest
          maximum: 20.0
          minimum: 1.0
          title: Max Topics
          type: integer
        min_confidence:
          default: 0.3
          description: Minimum confidence threshold for suggestions
          maximum: 1.0
          minimum: 0.0
          title: Min Confidence
          type: number
        store_result:
          default: true
          description: Whether to persist the classification result
          title: Store Result
          type: boolean
        taxonomy_hint:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Optional taxonomy hint to scope the classification
          title: Taxonomy Hint
      required:
      - content
      title: DocumentClassificationRequest
      type: object
    DocumentClassificationResponse:
      description: Complete classification response including taxonomy and topics.
      properties:
        alternative_taxonomies:
          items:
            $ref: '#/components/schemas/TaxonomySuggestion'
          title: Alternative Taxonomies
          type: array
        classification_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: ID of stored classification if persisted
          title: Classification Id
        content_preview:
          description: First 200 chars of content
          title: Content Preview
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        document_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Document Id
        document_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Type
        model_name:
          title: Model Name
          type: string
        model_version:
          title: Model Version
          type: string
        prompt_version:
          title: Prompt Version
          type: string
        suggested_taxonomy:
          $ref: '#/components/schemas/TaxonomySuggestion'
        suggested_topics:
          items:
            $ref: '#/components/schemas/TopicSuggestion'
          title: Suggested Topics
          type: array
      required:
      - content_preview
      - suggested_taxonomy
      - model_name
      - model_version
      - prompt_version
      - created_at
      title: DocumentClassificationResponse
      type: object
...
    FeedbackRequest:
      description: User feedback on a classification result.
      properties:
        classification_id:
          format: uuid
          title: Classification Id
          type: string
        corrected_taxonomy_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Corrected Taxonomy Id
        corrected_topic_ids:
          anyOf:
          - items:
              format: uuid
              type: string
            type: array
          - type: 'null'
          title: Corrected Topic Ids
        feedback:
          enum:
          - accepted
          - rejected
          - modified
          title: Feedback
          type: string
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      required:
      - classification_id
      - feedback
      title: FeedbackRequest
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    TaxonomySuggestion:
      description: A single taxonomy suggestion with confidence.
      properties:
        confidence:
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          type: number
        reasoning:
          anyOf:
          - type: string
          - type: 'null'
          title: Reasoning
        taxonomy_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Taxonomy Description
        taxonomy_id:
          format: uuid
          title: Taxonomy Id
          type: string
        taxonomy_identifier:
          title: Taxonomy Identifier
          type: string
        taxonomy_title:
          title: Taxonomy Title
          type: string
      required:
      - taxonomy_id
      - taxonomy_identifier
      - taxonomy_title
      - taxonomy_description
      - confidence
      title: TaxonomySuggestion
      type: object
...
    TopicSuggestion:
      description: A single topic suggestion with confidence and rank.
      properties:
        confidence:
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          type: number
        rank:
          minimum: 1.0
          title: Rank
          type: integer
        reasoning:
          anyOf:
          - type: string
          - type: 'null'
          title: Reasoning
        topic_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic Description
        topic_id:
          format: uuid
          title: Topic Id
          type: string
        topic_identifier:
          title: Topic Identifier
          type: string
        topic_title:
          title: Topic Title
          type: string
      required:
      - topic_id
      - topic_identifier
      - topic_title
      - topic_description
      - confidence
      - rank
      title: TopicSuggestion
      type: object
    TopicSuggestionRequest:
      description: Request payload for topic suggestions.
      properties:
        limit:
          default: 5
          description: Maximum number of suggestions to return.
          maximum: 20.0
          minimum: 1.0
          title: Limit
          type: integer
        min_confidence:
          default: 0.1
          description: Minimum confidence required to include a suggestion.
          maximum: 1.0
          minimum: 0.0
          title: Min Confidence
          type: number
        taxonomy_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Optional taxonomy scope for suggestions.
          title: Taxonomy Id
        text:
          description: Input text to classify.
          minLength: 1
          title: Text
          type: string
      required:
      - text
      title: TopicSuggestionRequest
      type: object
    TopicSuggestionResponse:
      description: Response payload for topic suggestions.
      properties:
        input_text:
          description: Original input text.
          title: Input Text
          type: string
        model_name:
          description: Classifier implementation name.
          title: Model Name
          type: string
        model_version:
          description: Classifier version.
          title: Model Version
          type: string
        suggestions:
          description: Ordered list of topic suggestions.
          items:
            $ref: '#/components/schemas/TopicSuggestionResult'
          title: Suggestions
          type: array
      required:
      - input_text
      - model_name
      - model_version
      title: TopicSuggestionResponse
      type: object
    TopicSuggestionResult:
      description: Single suggestion item returned to clients.
      properties:
        confidence:
          description: Confidence score.
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          type: number
        metadata:
          additionalProperties:
            items:
              type: string
            type: array
          description: Additional metadata for the suggestion.
          title: Metadata
          type: object
        rank:
          description: Rank order (1 = best).
          minimum: 1.0
          title: Rank
          type: integer
        slug:
          description: Topic slug.
          title: Slug
          type: string
        taxonomy_id:
          description: Taxonomy containing the topic.
          format: uuid
          title: Taxonomy Id
          type: string
        title:
          description: Topic title.
          title: Title
          type: string
        topic_id:
          description: Suggested topic ID.
          format: uuid
          title: Topic Id
          type: string
      required:
      - topic_id
      - taxonomy_id
      - title
      - slug
      - confidence
      - rank
      title: TopicSuggestionResult
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    email: mike@mikewaters.net
    name: LifeOS Development Team
    url: https://github.com/LifeOS/Substrate
  description: REST API for topic taxonomy and knowledge management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Ontology API
  version: 0.1.0
  x-api-id: ontology-api
  x-audience: internal
openapi: 3.1.0
paths:
...
  /classification/classify:
    post:
      description: "Classify document content into an appropriate taxonomy and suggest relevant topics.\n\n    This endpoint\
        \ performs a two-stage LLM-based classification:\n    1. Analyzes content to suggest which taxonomy it belongs to\n\
        \    2. Suggests relevant topics within the selected taxonomy\n\n    The classification results can optionally be\
        \ persisted for historical tracking."
      operationId: classify_document_classification_classify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentClassificationRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentClassificationResponse'
              example:
                content_preview: '# Dopamine and Neuroplasticity


                  Dopamine is a crucial neurotransmitter in the brain that plays a significant role in several

                  functions, including reward, motivation, memory, attention, and learning. I...'
                suggested_taxonomy:
                  taxonomy_id: 00000000-0000-0000-188d-6696711aee7d
                  taxonomy_identifier: tx:health
                  taxonomy_title: Health & Wellness
                  taxonomy_description: Physical and mental health topics
                  confidence: 0.95
                  reasoning: Document clearly focuses on neuroscience and brain health, discussing neurotransmitters, brain
                    function, and neurological conditions
                alternative_taxonomies: []
                suggested_topics:
                - topic_id: 00000000-0000-0000-1efa-920e8e68fd95
                  topic_identifier: health:neuroscience
                  topic_title: Neuroscience
                  topic_description: null
                  confidence: 0.92
                  rank: 1
                  reasoning: Core focus on neurotransmitter function and brain mechanisms
                - topic_id: 00000000-0000-0000-ba9e-f5ac06da5ea0
                  topic_identifier: health:mental
                  topic_title: Mental
                  topic_description: null
                  confidence: 0.78
                  rank: 2
                  reasoning: Discusses learning, memory, and attention which relate to mental health
                model_name: claude-sonnet-4-20250514
                model_version: 20250514
                prompt_version: 1.0.0
                classification_id: 00000000-0000-0000-9fde-8003bdd7f083
                created_at: '2025-11-04T02:55:11.302186Z'
                document_id: 00000000-0000-0000-9543-7e729e91ecbc
                document_type: Note
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Classify document into taxonomy and topics
      tags:
      - classification
  /classification/feedback:
    post:
      description: Record user feedback to improve future classifications.
      operationId: submit_feedback_classification_feedback_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Submit feedback on classification results
      tags:
      - classification
  /classification/history/{document_id}:
    get:
      description: Retrieve historical classification results for a document.
      operationId: get_classification_history_classification_history__document_id__get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          format: uuid
          title: Document Id
          type: string
      - description: Type of document
        in: query
        name: document_type
        required: true
        schema:
          description: Type of document
          title: Document Type
          type: string
      - description: Maximum results
        in: query
        name: limit
        required: false
        schema:
          default: 10
          description: Maximum results
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationHistoryResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get classification history for a document
      tags:
      - classification
  /classifier/suggestions:
    post:
      description: Return ranked topic suggestions for the supplied text.
      operationId: suggest_topics_classifier_suggestions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicSuggestionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicSuggestionResponse'
              example:
                input_text: Using Claude for building AI applications with transformers
                model_name: token_classifier
                model_version: 1.0.0
                suggestions:
                - topic_id: 00000000-0000-0000-df70-3781f1070368
                  taxonomy_id: 00000000-0000-0000-d315-cb6cbdc35142
                  title: Software
                  slug: software
                  confidence: 0.89
                  rank: 1
                  metadata: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Generate topic suggestions for input text
      tags:
      - classifier
...
servers:
- description: Development server
  url: http://localhost:8000
- description: Production server
  url: https://api.lifeos.dev